[DlgSettings] Refactor: split SettingsPage into separate files
