+++ cockatrice/resources/config/qtlogging.ini
@@ -28,6 +28,8 @@
 #dlg_tip_of_the_day = true
 #dlg_update = true
 
+#general_settings_page = true
+
 #settings_cache = true
 #servers_settings = true
 #shortcuts_settings = true
